Sampler -> workflow.Sampling move test sampling to test fixtures mover estimation sampling to SamplingForEstimation makePhantomCvrs -> audit package
