+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/estimate/TestMultiContestTestData.kt
@@ -13,6 +13,7 @@ import org.cryptobiotic.rlauxe.util.doubleIsClose
 import org.cryptobiotic.rlauxe.util.tabulateAuditableCards
 import org.cryptobiotic.rlauxe.util.tabulateVotesFromCvrs
 import org.cryptobiotic.rlauxe.verify.checkEquivilentVotes
+import org.cryptobiotic.rlauxe.workflow.ClcaWithoutReplacement
 import kotlin.math.abs
 import kotlin.test.Test
 import kotlin.test.assertEquals
