+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/alpha/CreateDiffMeansPolling.kt
@@ -26,7 +26,7 @@ import org.cryptobiotic.rlauxe.rlaplots.SRTcsvWriter
 import org.cryptobiotic.rlauxe.rlaplots.makeSRT
 import org.cryptobiotic.rlauxe.estimate.*
 import org.cryptobiotic.rlauxe.util.*
-import org.cryptobiotic.rlauxe.estimate.PollWithoutReplacement
+import org.cryptobiotic.rlauxe.workflow.PollWithoutReplacement
 import kotlin.test.Test
 
 // CANDIDATE FOR REFACTOR
