+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/corla/CorlaAudit.kt
@@ -120,7 +120,7 @@ class AuditCorlaAssertion(val quiet: Boolean = true): ClcaAssertionAuditorIF {
         auditConfig: AuditConfig,
         contestRound: ContestRound,
         assertionRound: AssertionRound,
-        sampler: Sampler,
+        sampling: Sampling,
         roundIdx: Int,
     ): TestH0Result {
         val contestUA = contestRound.contestUA
@@ -138,7 +138,7 @@ class AuditCorlaAssertion(val quiet: Boolean = true): ClcaAssertionAuditorIF {
             p1 = 0.0, p2 = 0.0, p3 = 0.0, p4 = 0.0, // TODO
         )
 
-        val testH0Result = testFn.testH0(sampler.maxSamples(), terminateOnNullReject = true) { sampler.sample() }
+        val testH0Result = testFn.testH0(sampling.maxSamples(), terminateOnNullReject = true) { sampling.sample() }
         val samplesNeeded = testH0Result.sampleCount
 
         val measuredCounts = if (testH0Result.tracker is ClcaErrorRatesIF)
@@ -148,8 +148,8 @@ class AuditCorlaAssertion(val quiet: Boolean = true): ClcaAssertionAuditorIF {
 
         assertionRound.auditResult = AuditRoundResult(
             roundIdx,
-            nmvrs = sampler.nmvrs(),
-            maxBallotIndexUsed = sampler.maxSampleIndexUsed(),
+            nmvrs = sampling.nmvrs(),
+            maxBallotIndexUsed = sampling.maxSampleIndexUsed(),
             pvalue = testH0Result.pvalueLast,
             samplesUsed = samplesNeeded,
             status = testH0Result.status,
